Make output_format matching more robust.
The issue at https://github.com/ruby-av/paperclip-av-transcoder/issues/3
is caused, at least on my system with Paperclip 4.2, by the format
parameter of the form: ".FMT".  This caused the case statement to not
match any of the desired formats!  It's likely that other paperclip
version don't pass the dot before the format extension, but it seems
likely that the behavior may differ depending on the version.

This change looks for the end of the format string to match using a
regex, and should now properly find the format, allowing thumbnails
to be generated (as well as the proper switches for non image params)