Fix failing Spec.
It was failing because it was manually adding the output_type,
while the Base#add_destination does it as well, causing the
options, aac and experimental to be added twice.

While this fixes the issue, I wonder if it might be better
to make the option setting idempontent?  Although with command-
line switches, I suppose one never knows when repeated arguments
are valid and necessary.